# src/server-web-next/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "r2f2 Website API",
    "version": "1.0.0",
    "description": "Browser uses HttpOnly cookies; App uses Supabase Bearer access token. All responses are private/no-store. Email is the login identifier; nicknames are display names only. Item, skill, set and order names use the browser r2f2-locale cookie or the App Accept-Language header. English and Chinese are sourced from game XML; Spanish falls back to English. The default is English."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
...
    "/game/accounts": {
      "get": {
        "summary": "List the authenticated player's game accounts",
        "operationId": "get_game_accounts",
        "security": [{ "cookieAuth": [] }, { "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Own accounts and character total levels", "content": { "application/json": { "schema": { "type": "object", "required": ["data"], "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/GameAccount" } } } } } } },
          "401": { "description": "UNAUTHENTICATED" },
          "403": { "description": "EMAIL_UNVERIFIED" },
          "502": { "description": "GAME_UNAVAILABLE" },
          "503": { "description": "GAME_UNAVAILABLE" }
        }
      },
      "post": {
        "summary": "Create a game account owned by the authenticated player",
        "operationId": "post_game_accounts",
        "description": "Browser writes require Origin equal to APP_ORIGIN. App writes require X-Client-Type: app and Bearer authentication. The owner email is derived from the verified session and cannot be supplied by the client.",
        "security": [{ "cookieAuth": [] }, { "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "additionalProperties": false, "required": ["name", "password", "passwordConfirmation"], "properties": { "name": { "type": "string", "minLength": 1, "maxLength": 10, "pattern": "^[!-~]+$" }, "password": { "type": "string", "minLength": 1, "maxLength": 10 }, "passwordConfirmation": { "type": "string" } } } } } },
        "responses": {
          "200": { "description": "Account created", "content": { "application/json": { "schema": { "type": "object", "required": ["data"], "properties": { "data": { "type": "object", "required": ["name"], "properties": { "name": { "type": "string" } } } } } } } },
          "400": { "description": "INVALID_INPUT" },
          "401": { "description": "UNAUTHENTICATED" },
          "403": { "description": "INVALID_ORIGIN / EMAIL_UNVERIFIED" },
          "409": { "description": "GAME_ACCOUNT_EXISTS" },
          "429": { "description": "RATE_LIMITED" },
          "502": { "description": "GAME_UNAVAILABLE" },
          "503": { "description": "GAME_UNAVAILABLE" }
        }
      }
    },
    "/auth/otp/send": {
      "post": {
        "summary": "Send signup/login email code",
        "operationId": "post_auth_otp_send",
        "description": "Signup may create a user. A verified existing email returns EMAIL_ALREADY_REGISTERED before sending mail; an unverified email may continue signup. Login never creates a user and does not disclose whether the address exists. Turnstile CAPTCHA is required.",
        "parameters": [
          {
            "in": "header",
            "name": "X-Client-Type",
            "schema": {
              "enum": ["browser", "app"],
              "default": "browser"
            },
            "description": "App clients must set app. App protected operations require Bearer Token and ignore cookies."
          },
          {
            "in": "header",
            "name": "Origin",
            "schema": {
              "type": "string"
            },
            "description": "Required and must equal APP_ORIGIN for browser write operations."
          }
        ],
        "security": [],
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "example": "private, no-store"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "sent": {
                          "const": true
                        },
                        "resendAfterSeconds": {
                          "const": 60
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "INVALID_INPUT / OTP_INVALID / AUTH_FAILED / CAPTCHA_REQUIRED / CAPTCHA_FAILED / WEBAUTHN_*",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "INVALID_ORIGIN / EMAIL_UNVERIFIED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "EMAIL_ALREADY_REGISTERED (signup only)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "PAYLOAD_TOO_LARGE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "UNSUPPORTED_MEDIA_TYPE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED (Retry-After: 60)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "502": {
            "description": "AUTH_UNAVAILABLE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "CONFIGURATION_REQUIRED / SERVICE_UNAVAILABLE / PROFILE_UNAVAILABLE / PASSKEY_UNAVAILABLE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "405": {
            "description": "METHOD_NOT_ALLOWED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpSend"
              }
            }
          }
        }
      }
    },
    "/auth/otp/verify": {
      "post": {
        "summary": "Verify email code and sign in",
        "operationId": "post_auth_otp_verify",
        "description": "",
        "parameters": [
          {
            "in": "header",
            "name": "X-Client-Type",
            "schema": {
              "enum": ["browser", "app"],
              "default": "browser"
            },
            "description": "App clients must set app. App protected operations require Bearer Token and ignore cookies."
          },
          {
            "in": "header",
            "name": "Origin",
            "schema": {
              "type": "string"
            },
            "description": "Required and must equal APP_ORIGIN for browser write operations."
          }
        ],
        "security": [],
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "example": "private, no-store"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": {
                      "oneOf": [
                        {
                          "$ref": "#/components/schemas/BrowserSession"
                        },
                        {
                          "$ref": "#/components/schemas/AppSession"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "INVALID_INPUT / OTP_INVALID / AUTH_FAILED / CAPTCHA_REQUIRED / CAPTCHA_FAILED / WEBAUTHN_*",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "INVALID_ORIGIN / EMAIL_UNVERIFIED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "NICKNAME_TAKEN / NICKNAME_COOLDOWN (includes nextNicknameChangeAt) / PROFILE_CHANGED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "PAYLOAD_TOO_LARGE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "UNSUPPORTED_MEDIA_TYPE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED (Retry-After: 60)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "502": {
            "description": "AUTH_UNAVAILABLE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "CONFIGURATION_REQUIRED / SERVICE_UNAVAILABLE / PROFILE_UNAVAILABLE / PASSKEY_UNAVAILABLE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "405": {
            "description": "METHOD_NOT_ALLOWED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpVerify"
              }
            }
          }
        }
      }
    },
    "/auth/passkeys/login/options": {
      "post": {
        "summary": "Start discoverable passkey sign-in",
        "operationId": "post_auth_passkeys_login_options",
        "description": "",
        "parameters": [
          {
            "in": "header",
            "name": "X-Client-Type",
            "schema": {
              "enum": ["browser", "app"],
              "default": "browser"
            },
            "description": "App clients must set app. App protected operations require Bearer Token and ignore cookies."
          },
          {
            "in": "header",
            "name": "Origin",
            "schema": {
              "type": "string"
            },
            "description": "Required and must equal APP_ORIGIN for browser write operations."
          }
        ],
        "security": [],
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "example": "private, no-store"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/PasskeyOptions"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "INVALID_INPUT / OTP_INVALID / AUTH_FAILED / CAPTCHA_REQUIRED / CAPTCHA_FAILED / WEBAUTHN_*",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "INVALID_ORIGIN / EMAIL_UNVERIFIED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "NICKNAME_TAKEN / NICKNAME_COOLDOWN (includes nextNicknameChangeAt) / PROFILE_CHANGED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "PAYLOAD_TOO_LARGE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "UNSUPPORTED_MEDIA_TYPE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED (Retry-After: 60)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "502": {
            "description": "AUTH_UNAVAILABLE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "CONFIGURATION_REQUIRED / SERVICE_UNAVAILABLE / PROFILE_UNAVAILABLE / PASSKEY_UNAVAILABLE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "405": {
            "description": "METHOD_NOT_ALLOWED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasskeyStart"
              }
            }
          }
        }
      }
    },
...
    "/auth/passkeys/register/options": {
      "post": {
        "summary": "Start passkey registration",
        "operationId": "post_auth_passkeys_register_options",
        "description": "",
        "parameters": [
          {
            "in": "header",
            "name": "X-Client-Type",
            "schema": {
              "enum": ["browser", "app"],
              "default": "browser"
            },
            "description": "App clients must set app. App protected operations require Bearer Token and ignore cookies."
          },
          {
            "in": "header",
            "name": "Origin",
            "schema": {
              "type": "string"
            },
            "description": "Required and must equal APP_ORIGIN for browser write operations."
          }
        ],
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "example": "private, no-store"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/PasskeyOptions"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "INVALID_INPUT / OTP_INVALID / AUTH_FAILED / CAPTCHA_REQUIRED / CAPTCHA_FAILED / WEBAUTHN_*",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "INVALID_ORIGIN / EMAIL_UNVERIFIED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "NICKNAME_TAKEN / NICKNAME_COOLDOWN (includes nextNicknameChangeAt) / PROFILE_CHANGED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "PAYLOAD_TOO_LARGE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "UNSUPPORTED_MEDIA_TYPE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED (Retry-After: 60)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "502": {
            "description": "AUTH_UNAVAILABLE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "CONFIGURATION_REQUIRED / SERVICE_UNAVAILABLE / PROFILE_UNAVAILABLE / PASSKEY_UNAVAILABLE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "405": {
            "description": "METHOD_NOT_ALLOWED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GameAccount": {
        "type": "object",
        "required": ["name", "characters"],
        "properties": {
          "name": { "type": "string" },
          "characters": { "type": "array", "items": { "type": "object", "required": ["name", "level"], "properties": { "name": { "type": "string" }, "level": { "type": "integer" } } } }
        }
      },
...
      "BrowserSession": {
        "type": "object",
        "required": ["user", "expiresAt"],
        "properties": {
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "expiresAt": {
            "type": "integer"
          }
        },
        "additionalProperties": false
      },
...
      "OtpSend": {
        "type": "object",
        "additionalProperties": false,
        "required": ["email", "intent"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 254
          },
          "intent": {
            "enum": ["signup", "login"]
          },
          "captchaToken": {
            "type": "string",
            "maxLength": 4096
          }
        }
      },
      "OtpVerify": {
        "type": "object",
        "additionalProperties": false,
        "required": ["email", "code"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "code": {
            "type": "string",
            "pattern": "^[0-9]{6}$"
          }
        }
      },
      "PasskeyStart": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "captchaToken": {
            "type": "string",
            "maxLength": 4096
          }
        }
      },
...
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message"],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "nextNicknameChangeAt": {
                "type": "string",
                "format": "date-time"
              }
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "r2f2_access"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}